min fix

+++ elevator_app/models.py
@@ -44,7 +44,7 @@ class Elevator(models.Model):
     building = models.ForeignKey(Building, on_delete=models.CASCADE)
 
     number = models.IntegerField()
-    current_floor = models.PositiveSmallIntegerField(default=0)
+    current_floor = models.IntegerField(default=0)
     is_busy = models.BooleanField(default=False)
     is_operational = models.BooleanField(default=True)
     is_door_open = models.BooleanField(default=False)
